refactor: move config to common dir
