+++ src/controllers/execute/sendAll.ts
@@ -2,7 +2,7 @@ import { userRepository } from '../../modules/db/repositories';
 import { rabbit, localizator } from '../../modules';
 import { SEND_ALL_SUBCOMMANDS, SEND_ALL_SUBCOMMANDS_LIST } from './constants';
 import CustomUserError from '../../modules/errors/UserError';
-import { app } from '../../config';
+import { app } from '../../common/config';
 
 export interface ExtraData {
   stickers?: string[];
